+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## fixed
 * `Console.blit` not taking source console alpha value into account
 * fixed #19 : rendering unicode characters >= 0x00FF
+* fixed #10 : console not centered in fullscreen and wrong mouse coordinates
 ## features
 * `AppOptions` now implements `Default`
 * added `InputApi.keys_released()` and `InputApi.keys_pressed()` that return iterators on key events since last update